Point TabManager at protocol
