+++ Tests/ClientTests/DependencyInjection/DependencyHelperMock.swift
@@ -17,7 +17,7 @@ class DependencyHelperMock {
         )
         AppContainer.shared.register(service: profile)
 
-        let tabManager: TabManager = TabManager(
+        let tabManager: TabManager = LegacyTabManager(
             profile: profile,
             imageStore: DiskImageStore(
                 files: profile.files,
