+++ Tests/ClientTests/Frontend/Browser/TabManagement/TabsQuantityTelemetryTests.swift
@@ -16,7 +16,7 @@ class TabsQuantityTelemetryTests: XCTestCase {
     }
 
     func testTrackTabsQuantity_withNormalTab_gleanIsCalled() {
-        let tabManager = TabManager(profile: MockProfile(), imageStore: nil)
+        let tabManager = LegacyTabManager(profile: MockProfile(), imageStore: nil)
         tabManager.addTab()
 
         TabsQuantityTelemetry.trackTabsQuantity(tabManager: tabManager)
@@ -31,7 +31,7 @@ class TabsQuantityTelemetryTests: XCTestCase {
     }
 
     func testTrackTabsQuantity_withPrivateTab_gleanIsCalled() {
-        let tabManager = TabManager(profile: MockProfile(), imageStore: nil)
+        let tabManager = LegacyTabManager(profile: MockProfile(), imageStore: nil)
         tabManager.addTab(isPrivate: true)
 
         TabsQuantityTelemetry.trackTabsQuantity(tabManager: tabManager)
