+++ Tests/ClientTests/HistoryHighlights/HistoryHighlightsManagerTests.swift
@@ -19,7 +19,7 @@ class HistoryHighlightsTests: XCTestCase {
         manager = HistoryHighlightsManager()
         profile = MockProfile(databasePrefix: "historyHighlights_tests")
         profile.reopen()
-        let tabManager = TabManager(profile: profile, imageStore: nil)
+        let tabManager = LegacyTabManager(profile: profile, imageStore: nil)
         entryProvider = HistoryHighlightsTestEntryProvider(with: profile, and: tabManager)
     }
