+++ Tests/ClientTests/StartAtHomeHelperTests.swift
@@ -16,7 +16,7 @@ class StartAtHomeHelperTests: XCTestCase {
 
         profile = MockProfile(databasePrefix: "startAtHomeHelper_tests")
         profile.reopen()
-        tabManager = TabManager(profile: profile, imageStore: nil)
+        tabManager = LegacyTabManager(profile: profile, imageStore: nil)
 
         FeatureFlagsManager.shared.initializeDeveloperFeatures(with: profile)
     }
